+++ packages/moengage_flutter/moengage_flutter_android/CHANGELOG.md
@@ -1,5 +1,12 @@
 # MoEngage Flutter Android Plugin
           
+# Next Release Date
+
+## Next Vesion
+- Add jsonEncode for missing APIs 
+- `moe-android-sdk` version updated to `13.05.01`
+- `inapp` version updated to `8.7.0`
+
 # 25-11-2024
 
 ## 3.1.1