+++ packages/moengage_inbox/moengage_inbox/CHANGELOG.md
@@ -1,5 +1,11 @@
 # MoEngage Flutter Inbox Plugin
           
+# Next Release Date
+
+## Next Release
+- Android
+  - `inbox-core` version updated to `3.2.1`
+
 # 25-11-2024
 
 ## 8.0.2